DEV Move BaseReader to serpentTools/parsers/base.py
