+++ serpentTools/parsers/history.py
@@ -9,7 +9,7 @@ file
 from numpy import empty, asfortranarray, array
 
 from serpentTools.utils import convertVariableName
-from serpentTools.objects.base import BaseReader
+from serpentTools.parsers.base import BaseReader
 from serpentTools.engines import KeywordParser
 from serpentTools.messages import warning, error
