+++ serpentTools/parsers/results.py
@@ -8,7 +8,7 @@ from numpy import array, vstack
 from serpentTools.settings import rc
 from serpentTools.utils import convertVariableName
 from serpentTools.objects.containers import HomogUniv
-from serpentTools.objects.base import XSReader
+from serpentTools.parsers.base import XSReader
 from serpentTools.utils import str2vec, splitValsUncs
 from serpentTools.messages import (warning, debug, SerpentToolsException)
